MSN Changes
MSN is probably now due for a total re-write. This patch changes MSN to process each stream separately, but tries to do them in order. This is a bit of a hack - the module should be re-written to account for this, and handle both streams. It also adds a number of basic tests that attempt to test basic MSN functionality. It also attempts to add Ver 15 support by basically ignoring a lot of the single sign on stuff. Some issues probably still exist.

darcs-hash:20071004043400-1ab04-a46cd39491f457155f2c8acf58dc82f58231650b.gz